requirements.txt
# requirements.txt
langchain
langchain-openai
langchain_experimental
langgraph  # Or another LLM provider library
-qU langchain-community unstructured
tqdm
fastapi
uvicorn[standard]
python-dotenv
chromadb
sentence-transformers # For embeddings
unstructured[pdf]   # For reading documents like 10-Ks
yfinance          # For the live stock price API tool
lxml              # Often needed by document loaders